+++ ceammc/ext/src/fx/fx_sdelay.cpp
@@ -9,4 +9,13 @@ EXTERNAL_NEW
     return p.pd_obj();
 }
 
-EXTERNAL_SETUP(fx);
+static void sdelay_clear(FAUST_EXT* x)
+{
+    x->dsp->instanceClear();
+}
+
+extern "C" void setup_fx0x2esdelay_tilde()
+{
+    internal_setup(gensym("fx.sdelay~"));
+    class_addmethod(FAUST_EXT_CLASS, (t_method)sdelay_clear, gensym("clear"), A_NULL);
+}